refactor: integrate beta changes
The more thorough beta test concluded very early on, that the NeoEnchant
provided enchants are too powerful and unfortunately they can not be
configured. This lead to the removal.

To compensate for some of the key enchants, additional mods have been
added as an replacement. See the changelog for the respective version
for more details.

Closes: #72